Add new App structure
